refactor(runner): extract scheduler event dispatch
